+++ src/tm.cpp
@@ -43,11 +43,11 @@ time::TimePoint compute_soft_limit(time::TimePoint               search_start,
     if (settings.w_time >= 0) {
 
         // Base time calculation
-        const auto compute_buffer_time = [&]() -> u64 {
+        const auto compute_buffer_time = [&]() -> f64 {
             if (stm == Color::White) {
-                return settings.w_time / 20 + settings.w_inc / 2;
+                return static_cast<f64>(settings.w_time / 20 + settings.w_inc / 2);
             } else {
-                return settings.b_time / 20 + settings.b_inc / 2;
+                return static_cast<f64>(settings.b_time / 20 + settings.b_inc / 2);
             }
         };
